+++ main.tf
@@ -8,7 +8,7 @@ module "cloud_provider_access" {
   iam_role_path                 = var.cloud_provider_access.iam_role_path
   iam_role_permissions_boundary = var.cloud_provider_access.iam_role_permissions_boundary
   tags                          = var.aws_tags
-  timeouts                      = var.timeouts.cloud_provider_access
+  timeouts                      = var.timeouts
 }
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -25,7 +25,7 @@ module "encryption_cloud_provider_access" {
   iam_role_path                 = var.encryption.iam_role.path
   iam_role_permissions_boundary = var.encryption.iam_role.permissions_boundary
   tags                          = var.aws_tags
-  timeouts                      = var.timeouts.cloud_provider_access
+  timeouts                      = var.timeouts
 }
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -47,6 +47,7 @@ module "encryption" {
   tags                       = var.aws_tags
   require_private_networking = local.encryption_require_private_networking
   enabled_for_search_nodes   = var.encryption.enabled_for_search_nodes
+  timeouts                   = var.timeouts
 
   depends_on = [module.cloud_provider_access, module.encryption_cloud_provider_access]
 }
@@ -57,7 +58,7 @@ module "encryption_private_endpoint" {
 
   project_id = var.project_id
   region     = each.key
-  timeouts   = var.timeouts.encryption_private_endpoint
+  timeouts   = var.timeouts
 
   depends_on = [module.encryption]
 }
@@ -76,7 +77,7 @@ module "backup_export_cloud_provider_access" {
   iam_role_path                 = var.backup_export.iam_role.path
   iam_role_permissions_boundary = var.backup_export.iam_role.permissions_boundary
   tags                          = var.aws_tags
-  timeouts                      = var.timeouts.cloud_provider_access
+  timeouts                      = var.timeouts
 }
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -94,6 +95,7 @@ module "backup_export" {
   bucket_name                 = var.backup_export.bucket_name
   create_s3_bucket            = var.backup_export.create_s3_bucket
   tags                        = var.aws_tags
+  timeouts                    = var.timeouts
 
   depends_on = [module.cloud_provider_access, module.backup_export_cloud_provider_access]
 }
@@ -112,7 +114,7 @@ module "log_integration_cloud_provider_access" {
   iam_role_path                 = var.log_integration.iam_role.path
   iam_role_permissions_boundary = var.log_integration.iam_role.permissions_boundary
   tags                          = var.aws_tags
-  timeouts                      = var.timeouts.cloud_provider_access
+  timeouts                      = var.timeouts
 }
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -133,6 +135,7 @@ module "log_integration" {
   kms_key                     = var.log_integration.kms_key
   kms_key_skip_iam_policy     = var.log_integration.kms_key_skip_iam_policy
   tags                        = merge(var.aws_tags, var.log_integration.tags)
+  timeouts                    = var.timeouts
 
   depends_on = [module.cloud_provider_access, module.log_integration_cloud_provider_access]
 }
@@ -147,7 +150,7 @@ resource "mongodbatlas_private_endpoint_regional_mode" "this" {
   enabled    = true
 
   dynamic "timeouts" {
-    for_each = var.timeouts.privatelink_regional_mode[*]
+    for_each = var.timeouts[*]
     content {
       create = timeouts.value.create
       update = timeouts.value.update
@@ -163,13 +166,12 @@ resource "mongodbatlas_privatelink_endpoint" "this" {
   region        = local._privatelink_aws_region[each.key]
 
   dynamic "timeouts" {
-    for_each = var.timeouts.privatelink_endpoint[*]
+    for_each = var.timeouts[*]
     content {
       create = timeouts.value.create
       delete = timeouts.value.delete
     }
   }
-  delete_on_create_timeout = try(var.timeouts.privatelink_endpoint.delete_on_create_timeout, null)
 }
 
 module "privatelink" {
@@ -198,7 +200,7 @@ module "privatelink" {
   }
 
   tags     = merge(var.aws_tags, each.value.tags)
-  timeouts = var.timeouts.privatelink_endpoint_service
+  timeouts = var.timeouts
 
   depends_on = [mongodbatlas_private_endpoint_regional_mode.this]
 }