+++ modules/privatelink/variables.tf
@@ -65,10 +65,11 @@ variable "tags" {
 
 variable "timeouts" {
   type = object({
-    create                   = optional(string)
-    delete                   = optional(string)
-    delete_on_create_timeout = optional(bool)
+    create = optional(string, "30m")
+    update = optional(string, "30m")
+    delete = optional(string, "30m")
   })
   default     = null
-  description = "Timeout overrides for privatelink_endpoint_service. See root module timeouts variable."
+  nullable    = true
+  description = "Timeout overrides. null = provider defaults."
 }